backend/requirements.txt
fastapi
uvicorn[standard]
boto3
python-multipart
pydantic
python-dotenv
